% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley}
\alias{barley}
\title{Barley microarray data}
\format{
A list containing raw_data, raw_meta, data (microarray data matrix with 104 samples, 22840 genes) and covariates (information of 104 samples).
}
\usage{
barley
}
\description{
This dataset contains gene expression measurements for barley (Hordeum vulgare) in response to the powdery mildew pathogen, as described in Caldo et al. (2004). The study investigates the interaction-dependent gene expression in the Mla-specified response to barley powdery mildew.
}
\references{
Caldo R A, Nettleton D, Wise R P. Interaction-dependent gene expression in Mla-specified response to barley powdery mildew\link{J}. The Plant Cell, 2004, 16(9): 2514-2528.
}
\keyword{barley}
