% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{Bootstrap Resampling of CEA Estimates}
\usage{
boot(
  x,
  R,
  weights = NULL,
  simple = FALSE,
  parallel = c("no", "multicore", "snow"),
  ncpus = getOption("cea.boot.ncpus", 1L),
  cl = NULL
)
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must use
the default 'formula' specification.}

\item{R}{The number of bootstrap replicates.}

\item{weights, simple, parallel, ncpus, cl}{Passed to \verb{\link[boot]\{boot\}}. For
\code{parallel} and \code{ncpus}, default values are taken from
\code{getOption(cea.boot.parallel)} and \code{getOption(cea.boot.ncpus)} instead
of their \code{boot}-package equivalents.}
}
\description{
Generate R bootstrap replicates of mean incremental QALYs and Costs from a
fitted CEA regression model, using the \code{boot} package. Only the ordinary
nonparametric bootstrap is currently implemented.
}
