% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea-package.R
\docType{package}
\name{cea-package}
\alias{cea}
\alias{cea-package}
\title{cea: Functions for Trial-Based Cost-Effectiveness Analysis}
\description{
Tools and functions to estimate and interpret trial-based
    cost-effectiveness analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uo-cmor/cea}
  \item Report bugs at \url{https://github.com/uo-cmor/cea/issues}
}

}
\author{
\strong{Maintainer}: Ross Wilson \email{ross.wilson@otago.ac.nz} (\href{https://orcid.org/0000-0001-8505-8081}{ORCID})

}
\keyword{internal}
