% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{ceac}
\alias{ceac}
\alias{ceac.cea_estimate}
\alias{ceac.cea_boot}
\title{Cost-Effectiveness Acceptability Curves}
\usage{
ceac(x, wtp_max, wtp_step, QALYs = "QALYs", Costs = "Costs", ...)

\method{ceac}{cea_estimate}(
  x,
  wtp_max,
  wtp_step,
  QALYs = "QALYs",
  Costs = "Costs",
  estimand = "ATE",
  method = "boot",
  R,
  sim = "ordinary",
  ...
)

\method{ceac}{cea_boot}(x, wtp_max, wtp_step, QALYs = "QALYs", Costs = "Costs", ...)
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must use
the default 'formula' specification.}

\item{wtp_max}{Maximum willingness-to-pay level for calculation.}

\item{wtp_step}{Interval between calculated willingness-to-pay levels.}

\item{QALYs, Costs}{Names of the variables in \code{x} representing QALYs and
Costs, respectively.}

\item{...}{Passed to \code{\link{boot}}.}

\item{estimand}{String scalar. Whether to calculate the average treatment
effect (ATE), average treatment effect on the treated (ATT), or average
treatment effect on the controls (ATC). Only used for non-linear models.}

\item{method}{Which method to use. Currently only 'boot' (bootstrap) is
implemented.}

\item{R}{The number of bootstrap replicates.}

\item{sim}{A character vector indicating the type of simulation required.
Possible values are "ordinary" (the default), "parametric", "balanced",
or "permutation".}
}
\description{
Generate a cost-effectiveness acceptability curve (CEAC) from a fitted CEA
regression model.
}
