% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{ci.cea_estimate}
\alias{ci.cea_boot}
\title{Confidence intervals for CEA estimates}
\usage{
ci(x, outcomes = "INMB", conf = 0.9, type = "bca", wtp, estimand = "ATE", ...)

\method{ci}{cea_estimate}(
  x,
  outcomes = "INMB",
  conf = 0.9,
  type = "bca",
  wtp,
  estimand = "ATE",
  method = "delta",
  R,
  sim = "parametric",
  ...
)

\method{ci}{cea_boot}(x, outcomes = "INMB", conf = 0.9, type = "bca", wtp, ...)
}
\arguments{
\item{x}{\code{cea_estimate} or \code{cea_boot} object. The fitted CEA regression
model or bootstrap resampling from the fitted model.}

\item{outcomes}{A character vector indicating the outcomes to be calculated.
Possible values are "QALYs", "Costs", "INMB", or "INHB".}

\item{conf}{Confidence level of the required intervals.}

\item{type}{Which type of intervals are required. Possible values are
"norm", "basic", "perc", or "bca".}

\item{wtp}{Willingness-to-pay level for calculation of INMB & INHB.}

\item{estimand}{String scalar. Whether to calculate the average treatment
effect (ATE), average treatment effect on the treated (ATT), or average
treatment effect on the controls (ATC). Only used for non-linear models.}

\item{...}{Passed to \code{\link{boot_cea}}.}

\item{method}{Which method to use. Available methods are 'boot' (bootstrap)
and 'delta' (delta approximation) are implemented.}

\item{R}{The number of bootstrap replicates.}

\item{sim}{A character vector indicating the type of simulation required.
Possible values are "ordinary" (the default), "parametric", "balanced",
or "permutation".}
}
\description{
Generate confidence intervals for mean incremental QALYs, costs, net
monetary benefit (INMB), and net health benefit (INHB) from a fitted CEA
regression model.
}
\details{
The 'boot' method uses \code{boot} and \code{boot.ci} from the \code{boot} package. The
'delta' method uses the delta approximation (\cite{Oehlert 1992}).
}
\references{
Oehlert GW. \emph{A note on the delta method}. Am Stat 1992;46(1):27-9.
\url{https://doi.org/10.2307/2684406}
}
