% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{Costs}
\alias{Costs}
\title{Incremental Intervention Costs}
\usage{
Costs(x, estimand = "ATE")
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must use
the default 'formula' specification.}

\item{estimand}{String scalar. Whether to calculate INMB from the average
treatment effect (ATE), average treatment effect on the treated (ATT),
or average treatment effect on the controls (ATC).}
}
\description{
Extract incremental costs from a fitted \code{cea_estimate} regression object.
}
\seealso{
Other treatment effect extractors: 
\code{\link{ICER}()},
\code{\link{INMB}()},
\code{\link{QALYs}()}
}
\concept{treatment effect extractors}
