% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{COVID-19 multi-omics data}
\format{
A list containing element metabolite (metabolite data matrix with 127 samples, 111 variables), protein (protein data matrix with 127 samples, 516 variables) and covariates (meta information).
}
\usage{
covid
}
\description{
A dataset including metabolome and proteome data of 127 patients in Albany Medical Center, NY from 6 April 2020 to 1 May 2020
}
\references{
Overmyer K A, Shishkova E, Miller I J, et al. Large-scale multi-omic analysis of COVID-19 severity\link{J}. Cell systems, 2021, 12(1): 23-40. e7.
}
\keyword{covid}
