% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPCA.R
\name{cPCA}
\alias{cPCA}
\title{Contrastive PCA}
\usage{
cPCA(Xt, Xc, eta, sparse_ctst = NULL, sparse_trt = NULL, sparse_ctrl = NULL)
}
\arguments{
\item{Xt}{data frame for treatment group, samples in rows, variables in columns.}

\item{Xc}{data frame for control group, samples in rows, variables in columns.}

\item{eta}{tuning parameter controlling how much control variation should be contrasted off from the treatment variation. It can be theo \code{eta_opt} from function eta_tuning()}

\item{sparse_ctst}{parameter controlling the sparsity of scPCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{sparse_trt}{parameter controlling the sparsity of treatment sparse PCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{sparse_ctrl}{parameter controlling the sparsity of control sparse PCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}
}
\value{
A list containing contrastive scores \code{cPC} and loadings \code{cU}, treatment scores \code{PCt} and loadings \code{Ut}, and control scores \code{PCc} and loadings \code{Uc}.
}
\description{
This function performs contrastive PCA (cPCA) or sparse contrastive PCA (scPCA), in comparison, also perform treatment/control group only PCA/sPCA
}
\examples{
library(CEA)
data(intro)
cpca_res <- cPCA(Xt = intro$cPCA$Xt, Xc = intro$cPCA$Xc, eta = 1.2, sparse_ctst = NULL, sparse_trt = NULL, sparse_ctrl = NULL)
}
