% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate joint model for intervention costs and effects (QALYs)}
\usage{
estimate(
  QALYs,
  costs,
  treatment,
  covars,
  data,
  linear_pred = NULL,
  matrix_pred = NULL,
  link = NULL,
  variance = NULL,
  ...
)
}
\arguments{
\item{QALYs, costs, treatment}{Character strings naming the variables in
\code{data} representing QALYs, costs, and treatment assignment,
respectively.}

\item{covars}{(optional) Character vector naming variables in \code{data}
included as (baseline) covariates in the regression models.}

\item{data}{A data frame (or object coercible by
\code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the model.}

\item{linear_pred}{(optional) A list of formula specifying the different
model components to be estimated. If specified, over-rides the model
specification in \code{QALYs}, \code{costs}, \code{treatment} and optional \code{covars}.
See \code{\link[mcglm]{mcglm}} for details (if provided, this argument
is passed to \code{mcglm} unchanged).}

\item{matrix_pred}{(optional) A list of matrices to be used on the matrix
linear predictor. See \code{\link[mcglm]{mcglm}} for details (if
provided, this argument is passed to \code{mcglm} unchanged). If not
specified, the default is to use an identity matrix for both components.}

\item{link}{(optional) A list of link function names. See
\code{\link[mcglm]{mcglm}} for details (if provided, this argument is
passed to \code{mcglm} unchanged). If not specified, the default is to use an
"identity" link for QALYs and a "log" link for Costs.}

\item{variance}{(optional) A list of link function names. See
\code{\link[mcglm]{mcglm}} for details (if provided, this argument is
passed to \code{mcglm} unchanged). If not specified, the default is to use a
"constant" variance function for QALYs and a "tweedie" variation
function for Costs.}

\item{...}{Optional arguments to be passed to \code{\link[mcglm]{mcglm}}.}
}
\description{
Estimate a multivariate covariance generalised linear model (\cite{Bonat &
Jørgensen 2016}) for the joint incremental costs and QALYs from a
randomised trial (or other comparable data source).
}
\details{
The standard model specification provides \code{QALYs}, \code{costs}, \code{treatment}, and
(optional) \code{covars}, to estimate a bivariate regression with common RHS
variables. If a different specification is required (e.g., estimating
additional outcomes beyond QALYs and Costs, or using different RHS
variables in each equation), a custom model specification can be passed
to \code{linear_pred} (which will be passed directly to
\code{\link[mcglm]{mcglm}}; see the \code{mcglm} documentation for
details of the model specification in this case).

In either specification, arguments \code{matrix_pred}, \code{link}, and
\code{variance} can optionally be specified to override the
corresponding default values.  See the documentation for
\code{\link[mcglm]{mcglm}} for the specification of these arguments.
If not specified, \code{matrix_pred} defaults to an identity matrix (for
each response variable); \code{link} defaults to \code{"identity"} for
QALYs and \code{"log"} for Costs if the default formula specification is
used, or \code{"identity"} for all response variables if a custom
\code{linear_pred} specification is provided; and \code{variance}
defaults to \code{"constant"} for QALYs and \code{"tweedie"} for Costs
(\code{"constant"} for all response variables if a custom
\code{linear_pred} specification is provided). These defaults correspond
to a linear regression model for QALYs and a Poisson-like GLM
(equivalent to \code{family = "quasipoisson"} in base \code{glm}) for
Costs, or a linear regression model for all responses for a custom
\code{"linear_pred"} specification.
}
\references{
Bonat WH, Jørgensen B. \emph{Multivariate covariance generalized linear
models}. J Royal Stat Soc 2016;65:649-75.
\url{https://doi.org/10.1111/rssc.12145}
}
