% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_tuning.R
\name{eta_tuning}
\alias{eta_tuning}
\title{Select \eqn{\eta} for subgroup identification by cPCA or scPCA}
\usage{
eta_tuning(
  Xt,
  Xc,
  eta_range = seq(0.5, 10, 0.5),
  plot = TRUE,
  sparse_trt = NULL,
  sparse_ctst = NULL,
  km_cluster = 2,
  ckm_cluster = 2
)
}
\arguments{
\item{Xt}{data frame for treatment group, samples in rows, variables in columns. If PCA is applied to correlation matrix, Xt should be scaled and centered.}

\item{Xc}{data frame for control group, samples in rows, variables in columns. If cPCA is applied to correlation matrix, Xc should also be scaled and centered.}

\item{eta_range}{a vector containing candidate eta range to be tuning from.}

\item{plot}{a logical quantity indicating whether plot the ARI(nuisance) and silhouette score.}

\item{sparse_trt}{parameter controlling the sparsity of treatment sparse PCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{sparse_ctst}{parameter controlling the sparsity of scPCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{km_cluster}{number showing number of nuisance clusters defined from kmeans clustering on treatment PCA}

\item{ckm_cluster}{number of interesting clusters defined from kmeans clustering on cPCA/scPCA}
}
\value{
A list containing selected value of \eqn{\eta}  \code{eta_opt}, the vector of ARI(nuisance) \code{nuisance_ARI} and the vector of Silhouette \code{silhouette}.
}
\description{
Select \eqn{\eta} for subgroup identification by cPCA or scPCA
}
