% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icer.R
\name{ICER}
\alias{ICER}
\title{Incremental Cost-Effectiveness Ratio}
\usage{
ICER(x, estimand = "ATE", ...)
}
\arguments{
\item{x}{\code{cea_estimate} object, fitted CEA regression model. Must use the
default 'formula' specification.}

\item{estimand}{String scalar. Whether to calculate ICER from the average
treatment effect (ATE), average treatment effect on the treated (ATT),
or average treatment effect on the controls (ATC).}

\item{...}{Not used.}
}
\description{
Calculate the incremental cost-effectiveness ratio (ICER) from a fitted
\code{cea_estimate} object.
}
