% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inb.R
\name{INMB}
\alias{INMB}
\alias{INHB}
\title{Incremental Net Monetary Benefit}
\usage{
INMB(x, wtp, estimand = "ATE", ...)

INHB(x, wtp, estimand = "ATE", ...)
}
\arguments{
\item{x}{\code{cea_estimate} object. Fitted CEA regression model. Must use the
default 'formula' specification.}

\item{wtp}{Numeric scalar.}

\item{estimand}{String scalar. Whether to calculate INMB from the average
treatment effect (ATE), average treatment effect on the treated (ATT),
or average treatment effect on the controls (ATC).}

\item{...}{Not used.}
}
\description{
Calculate the incremental net monetary benefit (INMB) or incremental net
health benefit (INHB) from a fitted \code{cea_estimate} object.
}
