% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intro}
\alias{intro}
\title{Read me data examples}
\format{
A list containing four sublists, each including data and covariates.
}
\usage{
intro
}
\description{
Read me data examples including simulated data for cPCA, 3CA and real barley and covid data for scPCA and s3CA.
}
\references{
Refer references for data(barley) and data(covid)
}
\keyword{intro}
