% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Acute myelogenous leukemia single cell sequencing data}
\format{
A list containing four single cell experiments object (sce)
}
\usage{
leukemia
}
\description{
scRNAseq for two healthy individuals (healthy 1 and 2), and two AML patients (AML027, 035) with top 1000 most variable genes. Pre-processing and gene selection is performed following Boileau 2020.
}
\references{
Boileau P, Hejazi N S, Dudoit S. Exploring high-dimensional biological data with sparse contrastive principal component analysis\link{J}. Bioinformatics, 2020, 36(11): 3422-3430. Zheng G X Y, Terry J M, Belgrader P, et al. Massively parallel digital transcriptional profiling of single cells\link{J}. Nature communications, 2017, 8(1): 14049.
}
\keyword{leukemia}
