% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_cea.R
\name{pool_cea}
\alias{pool_cea}
\title{Pool multiply-imputed CEA regression results}
\usage{
pool_cea(x, ...)
}
\arguments{
\item{x}{A \code{cea_mira} object. The fitted CEA regression models (on \code{m}
imputed datasets).}

\item{...}{For future expansion. Not currently used.}
}
\description{
Combine estimates from \code{m} repeated complete data CEA regression analyses,
using Rubin's pooling rules.
}
