% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qalys.R
\name{QALYs}
\alias{QALYs}
\title{Incremental Quality-Adjusted Life Year Gains}
\usage{
QALYs(x)
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must use
the default 'formula' specification.}
}
\description{
Extract incremental QALY gains from a fitted \code{cea_estimate} regression
object.
}
\seealso{
Other treatment effect extractors: 
\code{\link{Costs}()},
\code{\link{ICER}()},
\code{\link{INMB}()}
}
\concept{treatment effect extractors}
