% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_reg.R
\name{score_reg}
\alias{score_reg}
\title{Regression of continuous response on the score from multivariate analysis}
\usage{
score_reg(x, y, plot = FALSE, title = "")
}
\arguments{
\item{x}{score from multivariate analysis.}

\item{y}{continous external response.}

\item{plot}{whether or not plot the regression fit, default is FALSE.}

\item{title}{title for regression plot.}
}
\value{
regression fit, i.e., lm object.
}
\description{
This function performs regression of continuous response on the score from multivariate analysis like (s)3CA and (s)cPCA.
}
