% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data_CSC}
\alias{sim_data_CSC}
\title{Contrastive spectral clustering simulated data}
\format{
A list containing target sample Xt, ancillary sample Xa, target covariates covariates_t and ancillary covariates covariates_a.
}
\usage{
sim_data_CSC
}
\description{
Contrastive spectral clustering simulated data using n11 = 50, n12 = 25, n21 = 55, n22 = 30, m1tilde = 20, m2tilde = 30,
mu1 = 5, mu2 = 2, theta1 = -5, theta2 = 6, p1 = 5, p2 = 10, Sigma_t = Identity, Sigma_a = Identity, and set.seed(111)
}
\keyword{sim_data_CSC}
