% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_clust.R
\name{spec_clust}
\alias{spec_clust}
\title{Compute ARI of k means clustering and give groups}
\usage{
spec_clust(score_mat, group)
}
\arguments{
\item{score_mat}{score matrix from multivariate analysis like cPCA, samples in rows.}

\item{group}{the group member ship of the samples.}
}
\value{
ARI between the resulting clusters and the given groups.
}
\description{
This function performs K means clustering on the input score and computes the ARI between the resulting clusters and the given groups.
}
