% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.R
\name{ceac}
\alias{ceac}
\title{Cost-Effectiveness Acceptability Curves}
\usage{
ceac(x, method = "boot", R, sim = "ordinary", wtp_max, wtp_step, ...)
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must use
the default 'formula' specification.}

\item{method}{Which method to use. Currently only 'boot' (bootstrap) is
implemented.}

\item{R}{The number of bootstrap replicates.}

\item{sim}{A character vector indicating the type of simulation required.
Possible values are "ordinary" (the default), "parametric", "balanced",
or "permutation".}

\item{wtp_max}{Maximum willingness-to-pay level for calculation.}

\item{wtp_step}{Interval between calculated willingness-to-pay levels.}

\item{...}{Passed to \code{\link{boot}}.}
}
\description{
Generate a cost-effectiveness acceptability curve (CEAC) from a fitted CEA
regression model.
}
