% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence Intervals for CEA Estimates}
\usage{
ci(
  x,
  outcomes = "INMB",
  conf = 0.9,
  method = "boot",
  R,
  sim = "ordinary",
  type = "bca",
  wtp,
  ...
)
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must use
the default 'formula' specification.}

\item{outcomes}{A character vector indicating the outcomes to be calculated.
Possible values are "QALYs", "Costs", "INMB", or "INHB".}

\item{conf}{Confidence level of the required intervals.}

\item{method}{Which method to use. Currently only 'boot' (bootstrap) is
implemented.}

\item{R}{The number of bootstrap replicates.}

\item{sim}{A character vector indicating the type of simulation required.
Possible values are "ordinary" (the default), "parametric", "balanced",
or "permutation".}

\item{type}{Which type of intervals are required. Possible values are
"norm", "basic", "perc", or "bca".}

\item{wtp}{Willingness-to-pay level for calculation of INMB & INHB.}

\item{...}{Passed to \code{\link{boot}}.}
}
\description{
Generate confidence intervals for mean incremental QALYs, costs, net
monetary benefit (INMB), and net health benefit (INHB) from a fitted CEA
regression model.
}
