% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate joint regression model for intervention costs and effects (QALYs)}
\usage{
estimate(
  QALYs,
  costs,
  treatment,
  covars,
  data,
  centre = NULL,
  cluster = NULL,
  family = NULL,
  prior = NULL,
  method = "mcglm",
  fixed = NULL,
  random = NULL,
  verbose = FALSE,
  control = NULL,
  ...
)
}
\arguments{
\item{QALYs, costs, treatment}{Character strings naming the variables in
\code{data} representing QALYs, costs, and treatment assignment,
respectively.}

\item{covars}{(optional) Character vector naming variables in \code{data}
included as (baseline) covariates in the regression models.}

\item{data}{A data frame (or object coercible by
\code{\link[base]{as.data.frame}} to a data frame) or a \code{mids} object
containing the variables in the model.}

\item{centre}{(optional) Character string naming a variable in \code{data}
signifying centre membership in a multicentre trial.}

\item{cluster}{(optional) Character vector naming a variable in \code{data}
signifying cluster membership in a cluster-randomised trial. NOTE: At
present, \code{centre} and \code{cluster} do exactly the same thing (estimate a
mixed-effects model with centre/cluster random effects), and only one of
these arguments should be specified.}

\item{family}{(optional) A list of family functions specifying the
distribution and link function for each outcome variable. Each element
of the list should be a valid \code{family} value specified as for
\code{\link[stats]{glm}}. If not specified, the default is to use
\code{\link[stats]{gaussian}()} (i.e. OLS) for QALYs and
\code{\link[stats]{Gamma}("log")} for Costs.}

\item{prior}{For \code{method = 'brms'} only, specification of priors for
Bayesian model. See \code{\link[brms]{brm}} and
\code{\link[brms]{set_prior}} for details. If not specified, uses
\code{"normal(0, 5)"} for fixed effects coefficients and \code{brm} defaults for
all other parameters.}

\item{method}{\code{"mcglm"} or \code{"mglmmPQL"}, specifying which model
fitting algorithm to use.}

\item{fixed}{An alternative way to specify the 'fixed effects' component of
the model. Required if the RHS variables should differ for each equation
or outcomes other than QALYs and Costs are to be estimated. If used,
should be a (optionally named) list of two-sided model formulae
specifying the equations for each outcome measure, and will override any
values specified in \code{QALYs}, \code{costs}, and \code{covars}; the variable
specified in \code{treatment} must be included in the RHS of each formula. If
\code{fixed} does not have names, the LHS of each formula will be used as
component names.}

\item{random}{(optional) Specification of the random effects component of
the model. Corresponds to the \code{matrix_pred} argument of
\code{\link[mcglm]{mcglm}} and the \code{random} argument of
\code{\link[nlme]{lme}}; see the corresponding package help for details.
If not specified, the default is to include centre/cluster random
effects (intercepts) for each outcome (if applicable).}

\item{verbose}{Whether to print messages from the model fitting function
(default=\code{FALSE}).}

\item{control}{A list of arguments to be passed to the fitting algorithm
(argument \code{control_algorithm} of \code{\link[mcglm]{mcglm}} for method
\code{"mcglm"}; \code{control} (returned via the function
\code{\link[nlme]{lmeControl}}) of \code{\link[nlme]{lme}} for method
\code{"mglmmPQL"}).}

\item{...}{Further optional arguments to be passed to the underlying model
fitting function specified by \code{method}
(\code{\link[mcglm]{mcglm}}/\code{\link[nlme]{lme}}). See also the
'Details' section for additional considerations for specifying these
arguments.}
}
\value{
An object of class \code{cea_estimate} inheriting from \code{mcglm}.
}
\description{
Estimate a multivariate generalised linear model for the joint incremental
costs and QALYs from a randomised trial (or other comparable data
source). Models can be estimated as either a Multivariate Covariance
Generalized Linear Model (MCGLM; see
\cite{Bonat & J\/orgensen 2016}), using the \code{mcglm}
package (\cite{Bonat 2018}); or a Multivariate Generalized Linear Mixed
Model (MGLMM) via Penalized Quasi-Likelihood
(\cite{Achana et al. 2021}).
}
\details{
The standard model specification provides \code{QALYs}, \code{costs}, \code{treatment}, and
optionally \code{covars} and \code{centre}/\code{cluster}, to estimate a bivariate
regression with common RHS variables and optional clustering by trial
centre or clustered randomisation.

If a different specification is required (e.g., estimating additional
outcomes beyond QALYs and Costs, or using different RHS variables in
each equation), a custom model specification can be passed to
\code{fixed} as a list of formulae. In this case, \code{family} should also
be used to specify the GLM 'family' for each outcome.

Note the following details for model specification for each method:

\strong{mcglm:}

\itemize{
\item Model specification arguments \code{linear_pred}, \code{matrix_pred},
\code{link}, and \code{variance} are constructed via the \code{estimate()} arguments
\code{QALYs}, \code{costs}, \code{treatment}, and \code{covars}; \code{centre} or
\code{cluster}; and \code{family} (for which see below), and should not be
passed directly. For custom model specifications, use the arguments
\code{fixed}, \code{random}, and \code{family}.
\item Instead of a standard GLM family specification, \code{family} can also
be provided as a list where each element (corresponding to each model
outcome) is a named list with elements "family" (a name for the custom
family), "link", "variance" (passed as the corresponding arguments to
\code{\link[mcglm]{mcglm}}), and "power" (the (starting) power
parameter, currently only used for \code{variance = "tweedie"}). For the
"tweedie" variance function, a power value 0 corresponds to a gaussian
distribution, 1 to the (quasi-)Poisson distribution, 2 to the gamma
distribution, and 3 to the inverse gaussian. (As in \code{mcglm}, the dafault
is to fix the power parameter to its initial value.)
\item The \code{control_initial} argument is derived by fitting separate GLM
models to each outcome variable (as in the \code{mcglm} default) and cannot
be directly set by the user. The initial power parameter can be set via
\code{family} as descried above.
}

\strong{mglmmPQL:}

\itemize{
\item The \code{lme} argument \code{method} is renamed \code{nlme_method} to avoid a
clash with the same-named argument to \code{estimate}.
\item Formulae specified in \code{random} or \code{weights} (see below) can refer
to the reserved name \code{outvar}, an indicator variable identifying the
outcome variable to which an observation belongs in the transformed
long-form (stacked) data used in the underlying \code{lme} estimation.
\item By default, \code{weights} is set to
\code{nlme::varIdent(form = ~1 | outvar)}, to allow different variances for
each outcome variable. See \code{\link[nlme]{varIdent}} for details.
\item \code{max_iter} can be used to specify the maximum number of PQL
iterations. Use
\code{control =\link[nlme:lmeControl]{nlme::lmeControl}(maxIter = <x>)}
to set the maximum number of iterations for the \code{lme} optimization
algorithm.
}
}
\references{
Achana F, Gallacher D, Oppong R, et al. \emph{Multivariate generalized
linear mixed-effects models for the analysis of clinical trial-based
cost-effectiveness data}. Med Decis Making 2021;41(6):667-84.
\url{https://doi.org/10.1177/0272989X211003880}

Bonat WH. \emph{Multiple response variables regression models in R: The
mcglm package}. J Stat Soft 2018;84(4):1-30.
\url{https://doi.org/10.18637/jss.v084.i04}

Bonat WH, \enc{Jørgensen}{Jorgensen} B. \emph{Multivariate covariance
generalized linear models}. J Royal Stat Soc 2016;65:649-75.
\url{https://doi.org/10.1111/rssc.12145}
}
