% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract-effects}
\alias{extract-effects}
\alias{QALYs}
\alias{Costs}
\alias{ICER}
\alias{INMB}
\alias{INHB}
\alias{extract}
\title{Incremental treatment eFfects}
\usage{
QALYs(x, estimand = "ATE")

Costs(x, estimand = "ATE")

ICER(x, estimand = "ATE")

INMB(x, wtp, estimand = "ATE")

INHB(x, wtp, estimand = "ATE")

extract(x, outcome, estimand = "ATE")
}
\arguments{
\item{x}{\code{cea_estimate} object. The fitted CEA regression model. Must
contain a 'QALYs' equation.}

\item{estimand}{String scalar. Whether to calculate the average treatment
effect (ATE), average treatment effect on the treated (ATT), or average
treatment effect on the controls (ATC). Only used for non-linear models.}

\item{wtp}{Numeric scalar. Willingness-to-pay per QALY gained.}

\item{outcome}{String scalar. Name of the outcome measure to extract.}
}
\description{
Extract incremental treatment effects from a fitted \code{cea_estimate}
regression object.
}
